# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: F1 Draft League API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues:
    get:
      operationId: listLeagues
      tags: [leagues]
      summary: List public leagues (searchable)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: seasonId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"

    post:
      operationId: createLeague
      tags: [leagues]
      summary: Create a league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
...
  /leagues/{id}:
    get:
      operationId: getLeague
      tags: [leagues]
      summary: Get league by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          description: Not found

    patch:
      operationId: updateLeague
      tags: [leagues]
      summary: Update league (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"

    delete:
      operationId: deleteLeague
      tags: [leagues]
      summary: Delete league (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /leagues/{id}/join:
    post:
      operationId: joinLeague
      tags: [leagues]
      summary: Join a public league
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /leagues/{id}/leave:
    post:
      operationId: leaveLeague
      tags: [leagues]
      summary: Leave a league
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /leagues/join-by-invite/{inviteCode}:
    post:
      operationId: joinLeagueByInvite
      tags: [leagues]
      summary: Join a private league by invite code
      parameters:
        - name: inviteCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          description: Invalid invite code
  /leagues/{id}/standings:
    get:
      operationId: getLeagueRankings
      tags: [leagues]
      summary: Get league standings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueStandings"
  /leagues/{id}/members:
    get:
      operationId: getLeagueMembers
      tags: [leagues]
      summary: Get league members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeagueMember"
  /leagues/{id}/kick/{userId}:
    delete:
      operationId: kickLeagueMember
      tags: [leagues]
      summary: Kick member from league (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Kicked

  # FRIENDS
...
  /scores/global:
    get:
      operationId: getGlobalRanking
      tags: [scores]
      summary: Get global season ranking
      parameters:
        - name: seasonId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingEntry"
  /scores/me:
    get:
      operationId: getMyScoreSummary
      tags: [scores]
      summary: Get my score summary for current season
      parameters:
        - name: seasonId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreSummary"
  /scores/gp/{gpId}:
    get:
      operationId: getGPScoreboard
      tags: [scores]
      summary: Get scoreboard for a specific GP
      parameters:
        - name: gpId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingEntry"

  # ADMIN
...
components:
  schemas:
...
    League:
      type: object
      required: [id, name, isPublic, inviteCode, ownerId, seasonId, memberCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        isPublic:
          type: boolean
        inviteCode:
          type: string
        ownerId:
          type: string
        ownerHandle:
          type: string
        seasonId:
          type: string
        memberCount:
          type: integer
        isFactory:
          type: boolean
        createdAt:
          type: string
    LeagueInput:
      type: object
      required: [name, isPublic, seasonId]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        isPublic:
          type: boolean
        seasonId:
          type: string
    LeagueUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        isPublic:
          type: boolean
    LeagueStandings:
      type: object
      required: [leagueId, leagueName, seasonId, entries]
      properties:
        leagueId:
          type: string
        leagueName:
          type: string
        seasonId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/RankingEntry"
    LeagueMember:
      type: object
      required: [userId, handle, displayName, joinedAt]
      properties:
        userId:
          type: string
        handle:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        joinedAt:
          type: string
        isOwner:
          type: boolean

    # FRIENDS
...
    RankingEntry:
      type: object
      required: [rank, userId, handle, displayName, totalPoints]
      properties:
        rank:
          type: integer
        userId:
          type: string
        handle:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        totalPoints:
          type: number
        gpCount:
          type: integer
        bonusBudget:
          type: number
    ScoreSummary:
      type: object
      required: [userId, seasonId, totalPoints, rank, gpBreakdown]
      properties:
        userId:
          type: string
        seasonId:
          type: string
        totalPoints:
          type: number
        rank:
          type: integer
        totalUsers:
          type: integer
        budget:
          type: number
        bonusBudget:
          type: number
        gpBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/GPPointEntry"
    GPPointEntry:
      type: object
      required: [gpId, gpName, round, points]
      properties:
        gpId:
          type: string
        gpName:
          type: string
        round:
          type: integer
        points:
          type: number
        rank:
          type: integer

    # ADMIN
...